pyproject.toml
[project]
name = "neuland-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mistralai",
    "requests",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "jinja2",
    "python-multipart",
    "pydantic",
    "pytest",
    "httpx",
    "sentence-transformers",
    "chromadb",
    "numpy",
    "langchain-text-splitters",
    "docker",
    "langchain-mistralai>=0.2.12",
    "langchain>=0.3.27",
    "langchain-core>=0.3.76",
    "langgraph>=0.6.7",
]

[dependency-groups]
dev = [
    "chromadb>=1.1.0",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.76",
    "langgraph>=0.6.7",
    "mistralai>=1.9.10",
    "python-dotenv>=1.1.1",
]
